chore: 불필요한 형변환 코드 제거

+++ src/main/java/com/codeit/mople/domain/conversation/service/ConversationService.java
@@ -132,7 +132,7 @@ public class ConversationService {
         nextCursor,
         nextIdAfter,
         hasNext,
-        (long) conversationDtos.size(),
+        conversationDtos.size(),
         request.sortBy(),
         request.sortDirection()
     );
